Refactor exception hierarchy for TransactionHistoryService
- Introduced `TransactionHistoryBaseException` as the parent class for transaction-related exceptions.
- Moved and updated `UnsupportedExporterException` and `TransactionFilterException` to extend the new base exception.
- Adjusted imports across modules to reflect the updated exception package structure.
