+++ src/main/java/info/mackiewicz/bankapp/system/banking/history/exception/TransactionFilterException.java
@@ -1,11 +1,11 @@
-package info.mackiewicz.bankapp.presentation.exception;
+package info.mackiewicz.bankapp.system.banking.history.exception;
 
 import info.mackiewicz.bankapp.shared.core.error.ErrorCode;
 
 /**
  * Exception thrown when transaction filtering fails
  */
-public class TransactionFilterException extends PresentationBaseException {
+public class TransactionFilterException extends TransactionHistoryBaseException {
     
     private static final ErrorCode ERROR_CODE = ErrorCode.INTERNAL_ERROR;
