+++ src/main/java/info/mackiewicz/bankapp/system/banking/history/exception/UnsupportedExporterException.java
@@ -1,8 +1,8 @@
-package info.mackiewicz.bankapp.presentation.exception;
+package info.mackiewicz.bankapp.system.banking.history.exception;
 
 import info.mackiewicz.bankapp.shared.core.error.ErrorCode;
 
-public class UnsupportedExporterException extends PresentationBaseException {
+public class UnsupportedExporterException extends TransactionHistoryBaseException {
     
     private static final ErrorCode ERROR_CODE = ErrorCode.UNSUPPORTED_EXPORTER;
