+++ src/main/java/info/mackiewicz/bankapp/system/banking/history/service/TransactionFilterService.java
@@ -1,7 +1,7 @@
 package info.mackiewicz.bankapp.system.banking.history.service;
 
 import info.mackiewicz.bankapp.account.model.Account;
-import info.mackiewicz.bankapp.presentation.exception.TransactionFilterException;
+import info.mackiewicz.bankapp.system.banking.history.exception.TransactionFilterException;
 import info.mackiewicz.bankapp.transaction.model.Transaction;
 import info.mackiewicz.bankapp.transaction.model.TransactionStatus;
 import info.mackiewicz.bankapp.transaction.model.TransactionType;
